project/requirement.txt
Flask==2.0.1
Flask-WTF==0.15.1
Werkzeug==2.0.1
WTForms==2.3.3
Pillow==8.3.2
opencv-python==4.5.3.56
numpy==1.21.2
daltonlens==0.2.0
